+++ config/dev.exs
@@ -1,15 +1,5 @@
 import Config
 
-# Configure your database
-config :mr_eric, MrEric.Repo,
-  username: "postgres",
-  password: "postgres",
-  hostname: "localhost",
-  database: "mr_eric_dev",
-  stacktrace: true,
-  show_sensitive_data_on_connection_error: true,
-  pool_size: 10
-
 # For development, we disable any cache and enable
 # debugging and code reloading.
 #